pyproject.toml
[tool.poetry]
name = "pytezos"
version = "3.2.0"
description = "Python toolkit for Tezos"
license = "MIT"
authors = ["Michael Zaikin <mz@baking-bad.org>", "Arthur Breitman", "Roman Serikov"]
readme = "README.md"
repository = "https://github.com/baking-bad/pytezos"
homepage = "https://pytezos.org"
keywords = ['tezos', 'blockchain', 'sdk', 'michelson', 'repl', 'cryptocurrencies', 'smart-contracts', 'jupyter']
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: System :: Shells",
    "Framework :: IPython",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3"

]
packages = [
    { include = "pytezos", from = "src" },
    { include = "michelson_kernel", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.7"
pyblake2 = "^1.1.2"
base58 = "^1.0.3"
requests = "^2.21.0"
pendulum = "*"
loguru = "*"
netstruct = "*"
tqdm = "*"
ply = "*"
simplejson = "*"
mnemonic = "*"
pyyaml = "*"
pysha3 = "1.0.2"
pysodium = "0.7.7"
secp256k1 = "0.13.2"
fastecdsa = "1.7.5"
bson = "^0.5.10"
strict_rfc3339 = "0.7"
py_ecc = "*"
deprecation = "*"
testcontainers = "^3.2.0"
click = "^7.1.2"
typing-extensions = "^3.7.4"
tabulate = "^0.7.5"
ipykernel = "^5.5.0"
jupyter-client = "^6.1.12"
attrs = "^20.3.0"
notebook = "^6.3.0"
jsonschema = "^3.2.0"
cattrs = "^1.3.0"
cached-property = "^1.5.2"
cattrs-extras = "^0.1.1"
bravado = "^11.0.3"
docker = "^4.4.4"
pydantic = "^1.8.2"
"ruamel.yaml" = "^0.17.4"
fcache = "^0.4.7"
Jinja2 = "^3.0.1"

[tool.poetry.dev-dependencies]
parameterized = "^0.7.0"
sphinx = "*"
sphinx-rtd-theme = "^0.5.2"
sphinx-sitemap = "*"
sphinxcontrib-googleanalytics = "^0.1"
bump2version = "^1.0.1"
pylint = "^2.7.2"
pylint-exit = "^1.2.0"
isort = "^5.7.0"
mypy = "^0.812"
diff-cover = "^4.2.3"
black = "^20.8b1"
nose = "^1.3.7"
coverage = "^5.5"
nose-timer = "^1.0.1"
sphinx-click = "^2.7.1"
pytest = "^6.2.4"

[tool.poetry.scripts]
pytezos = 'pytezos.cli.cli:cli'
michelson-kernel = 'michelson_kernel.cli:cli'

[tool.isort]
line_length = 140

[tool.black]
line-length = 140
target-version = ['py37', 'py38']
skip-string-normalization = true

[build-system]
requires = ["poetry_core>=1.0.0", "cryptography==3.3.2", "wheel"]
build-backend = "poetry.core.masonry.api"
